requirements/base.txt
psycopg2==2.9.9

celery==5.4.0
django-celery-results==2.5.1
django-celery-beat==2.6.0

whitenoise==6.7.0

boto3==1.34.144
attrs==23.2.0

gunicorn==22.0.0
sentry-sdk==1.40.5

requests==2.32.3

google-api-python-client==2.86.0
google-auth==2.21.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0

pyotp==2.9.0
qrcode==7.4.2

# Django
# ------------------------------------------------------------------------------
Django==5.0.7
django-ninja==1.2.2
django-ninja-jwt==5.3.2
django-ninja-extra==0.21.2 # Used by JWT to register controllers to be used as auth mechanism
djangorestframework==3.15.2 # TODO STACK: Replace with Django Ninja
drf-jwt==1.19.2
django-environ==0.11.2
django-filter==24.2
django-extensions==3.2.3
django-cors-headers==4.4.0
django-storages==1.14.4
